Create script
This script will first wait for 1000 milliseconds (1 second) to allow the Android device to fully boot up and be ready for input. It will then type each of the 5 common PIN codes one by one, followed by an ENTER key press to submit the code. After each code is entered, there is a delay of 500 milliseconds to allow the device to process the input before the next code is entered.